Added tennessee-example-irv-tie.abif and streamlined winnerstr output
Previously, the IRV winnerstr was calculated in the template, which
was kinda silly.  Now that IRV_dict['winnerstr'] is available in the
abiflib pairwise json output, it can be used in awt.  Also, the tie
contest was added to abiftool to test what happens when there's an IRV
tie election.