Fixing last test in test_awt_routes.py
I added test_awt_routes.py in an earlier commit (dd632b824 - "Bumping
version to 0.31.0"), but I neglected to note the most significant
part: a test suite for awt!  Now, not only is there a test suite for
awt, but all of the tests pass (at least, on my machine).