Many fixes in advance of uploading awt 0.31.0rc2 to pypi
I relied on guidance from cursor.com after Google Gemini's tooling hit
a wall.  Gemini had a lot to do with the last commit (and other stuff
I've done in this codebase), but it wasn't willing to work around my
funky file layout in the awt codebase.  Cursor was incredibly helpful
in this stage of work.