feat(html): Using the full candidate name on the results page
Previously, the candidates were referred to by their candidate id
token in many places on results-index.html.  Most of this patch was
worked up by Claude Sonnet 4.  There's a lot of logic inside the
templates that really should go into the application controller, but
I'm not in the mood to be fussy right now.