feat(irv): Adding front-end display to abiftool back-end IRVextra information
The goal of the IRVextra branches in awt and abiftool is to fix awt
issue #10 in the github git repo:
https://github.com/electorama/awt/issues/10

In short, in the IRV result tables, it would be helpful to see the
second preferences for voters to get a general sense of the voter
sentiment for a batch of voters. This particularly matters for the
penultimate round and final round of voting, since the folks who vote
for the runner-up candidate in the final round don't get their second
choices considered at all.