feat(html): made ToC into a table with cand colorboxes; template robustification
I don't really like the table of contents as a table, but I like
having the colorboxes line up.  While I was testing the changes Claude
made to fix a lot of the tech debt around colorboxes, I discovered
that unchecking an election method caused server-side exceptions.
There's some abiflib backend changes we need to make before futzing
with awt some more, but for now, it's good enough for me to use.