test(html): tests for homepage tab failure
Adding tests to detect a problem with the homepage.  I only discovered
a bug where the homepage showed five textareas unlike the one that it
was supposed to. The original design called for showing only one
textarea at a time (keeping the other four hidden).  That stopped
working after commit 9341751b.  It worked with aa8539ff (the commit
just prior).  The new tests should detect this problem.